refactor: formatter related
